feat(connector): remove doris-catalog module,  doris module add catalog
