--- datavines-connector/datavines-connector-plugins/datavines-connector-doris-catalog/src/main/java/io/datavines/connector/plugin/DorisCatalogExecutor.java
@@ -1,39 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.datavines.connector.plugin;
-
-import io.datavines.common.datasource.jdbc.BaseJdbcDataSourceInfo;
-import io.datavines.connector.api.DataSourceClient;
-
-import java.util.Map;
-
-/**
- * Doris Catalog Executor
- * 执行器，负责执行 SQL 语句和查询操作
- * SWITCH CATALOG 逻辑在 DorisCatalogConnector 中实现
- */
-public class DorisCatalogExecutor extends DorisExecutor {
-
-    public DorisCatalogExecutor(DataSourceClient dataSourceClient) {
-        super(dataSourceClient);
-    }
-
-    @Override
-    public BaseJdbcDataSourceInfo getDatasourceInfo(Map<String,String> paramMap) {
-        return new DorisCatalogDataSourceInfo(paramMap);
-    }
-}
