--- datavines-connector/datavines-connector-plugins/datavines-connector-doris-catalog/src/main/resources/META-INF/plugins/io.datavines.connector.api.ConnectorFactory
@@ -1,1 +1,0 @@
-io.datavines.connector.plugin.DorisCatalogConnectorFactory
