+++ datavines-connector/datavines-connector-plugins/pom.xml
@@ -43,7 +43,6 @@
         <module>datavines-connector-trino</module>
         <module>datavines-connector-starrocks</module>
         <module>datavines-connector-doris</module>
-        <module>datavines-connector-doris-catalog</module>
         <module>datavines-connector-databend</module>
         <module>datavines-connector-mongodb</module>
         <module>datavines-connector-sqlserver</module>
